pyproject.toml
[project]
name = "mcp-email-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.164.0",
    "google-auth-oauthlib>=1.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.0",
]
